requirements.txt
streamlit
langchain-google-genai
langchain
PyPDF2
python-docx